+++ code/include/MergePolicy.hpp
@@ -5,8 +5,8 @@
  *        and a function for recombining solutions.
  * 
  * @author Francesco Biscaccia Carrara
- * @version v1.0.6
- * @since 04/04/2025
+ * @version v1.1.0
+ * @since 04/08/2025
  */
 
 #ifndef MER_POL_H