+++ code/source/FixPolicy.cpp
@@ -8,14 +8,14 @@ bool isInteger(double n) {
 	return ((n < CPX_INFBOUND) && (static_cast<int>(n) == n));
 }
 
-void FixPolicy::firstThetaFixing(std::vector<double>& x,std::string fileName, double theta, Random rnd) {
+void FixPolicy::startSolTheta(std::vector<double>& sol,std::string fileName, double theta, Random& rnd) {
 	if (theta < EPSILON || theta > 1.0)
 		throw FixPolicyException(FPEx::InputSizeError, "Theta par. must be within (0,1)!");
 
 
 	RlxFMIP	 relaxedFMIP{ fileName };
 	int		 numVarsToFix{ relaxedFMIP.getMIPNumVars() };
-	x.resize(numVarsToFix,CPX_INFBOUND);	
+	sol.resize(numVarsToFix,CPX_INFBOUND);	
 
 	
 	std::vector<size_t>					varRangesIndices(numVarsToFix);
@@ -45,15 +45,15 @@ void FixPolicy::firstThetaFixing(std::vector<double>& x,std::string fileName, do
 				double clampedLower = std::max(-MAX_UB, lowerBound);
 				double clampedUpper = std::min(MAX_UB, upperBound);
 
-				x[idx] = rnd.Int(clampedLower, clampedUpper);
-				relaxedFMIP.setVarValue(idx,x[idx]);
+				sol[idx] = rnd.Int(clampedLower, clampedUpper);
+				relaxedFMIP.setVarValue(idx,sol[idx]);
 				isFixed[idx] = true;
 				fixedThisIteration++;
 			}
 		}
 
 #if ACS_VERBOSE >= VERBOSE
-		PRINT_INFO("FixPolicy::firstThetaFixing - %zu vars hard-fixed", varsToFix);
+		PRINT_INFO("FixPolicy::startSolTheta - %zu vars hard-fixed", varsToFix);
 #endif
 
 		relaxedFMIP.solveRelaxation();
@@ -62,8 +62,8 @@ void FixPolicy::firstThetaFixing(std::vector<double>& x,std::string fileName, do
 		for (size_t i = 0; i < numVarsToFix; ++i) {
 			if(isFixed[i]) continue;
 			if(isInteger(lpSol[i])) {
-				x[i] = lpSol[i];
-				relaxedFMIP.setVarValue(i,x[i]);
+				sol[i] = lpSol[i];
+				relaxedFMIP.setVarValue(i,sol[i]);
 				isFixed[i] = true;
 			}
 		}
@@ -73,7 +73,7 @@ void FixPolicy::firstThetaFixing(std::vector<double>& x,std::string fileName, do
 
 }
 
-void FixPolicy::randomRhoFix(const std::vector<double>& sol, MIP& model, const size_t threadID, double rho,const char* type, Random& rnd) {
+void FixPolicy::randomRhoFixMT(const size_t threadID, const char* type, MIP& model, const std::vector<double>& sol, double rho, Random& rnd) {
 	if (rho < EPSILON || rho >= 1.0)
 		throw FixPolicyException(FPEx::InputSizeError, "Rho par. must be within (0,1)!");
 
@@ -82,11 +82,97 @@ void FixPolicy::randomRhoFix(const std::vector<double>& sol, MIP& model, const s
 	const size_t start = rnd.Int(0,xLen - 1);
 
 #if ACS_VERBOSE >= VERBOSE
-	PRINT_INFO("Proc: %3d [%s] - FixPolicy::randomRhoFix - %zu vars hard-fixed", threadID, type, numFixedVars);
+	PRINT_INFO("Proc: %3d [%s] - FixPolicy::randomRhoFixMT - %zu vars hard-fixed [%5.4f]", threadID, type, numFixedVars,rho);
 #endif
 
 	for (size_t i{ 0 }; i < numFixedVars; i++) {
 		size_t index{ (start + i) % xLen};
 		model.setVarValue(index, sol[index]);
 	}
+}
+
+
+void FixPolicy::dynamicAdjustRho(const char* phase, const int solveCode, const size_t numMIPs, double& CLIRho, const size_t A_RhoChanges){
+	if( A_RhoChanges >= numMIPs) return;
+
+	double scaledDeltaRho = 2*DELTA_RHO/numMIPs;
+
+	switch(solveCode){
+		case CPXMIP_OPTIMAL:
+		case CPXMIP_OPTIMAL_TOL:
+			CLIRho = (CLIRho-scaledDeltaRho < MIN_RHO)? MIN_RHO : CLIRho-scaledDeltaRho;
+#if ACS_VERBOSE >= VERBOSE
+			PRINT_INFO("[%s] - FixPolicy::dynamicAdjustRho - Rho Decreased [%5.4f]", phase, CLIRho);
+#endif
+			break;
+
+
+		case CPXMIP_DETTIME_LIM_FEAS:
+		case CPXMIP_TIME_LIM_FEAS:
+			CLIRho = (CLIRho+scaledDeltaRho > MAX_RHO)? MAX_RHO : CLIRho+scaledDeltaRho;
+#if ACS_VERBOSE >= VERBOSE
+			PRINT_INFO("[%s] - FixPolicy::dynamicAdjustRho - Rho Increased [%5.4f]", phase, CLIRho);
+#endif
+			break;
+
+		default: 
+#if ACS_VERBOSE >= VERBOSE
+		PRINT_ERR("Unexpected value for solvecode: %d",solveCode);
+#endif
+
+			break;
+	}
+}
+
+
+
+//Values to count the number of decrements/increments of Rho parameter in 1 iteration
+static size_t  numDecRho=0;
+static size_t  numIncRho=0;
+
+void FixPolicy::dynamicAdjustRhoMT(const size_t threadID,const char* type, const int solveCode,const size_t numMIPs, double& CLIRho, std::atomic_size_t& A_RhoChanges){
+	
+	if( A_RhoChanges >= numMIPs) return; // Whenever an incumbent is found, we don't touch Rho anymore
+	if( A_RhoChanges == 0 ) numDecRho = numIncRho = 0;
+
+	double scaledDeltaRho = DELTA_RHO/numMIPs;
+	
+	switch(solveCode){
+		case CPXMIP_OPTIMAL:
+		case CPXMIP_OPTIMAL_TOL:
+			CLIRho = (CLIRho-scaledDeltaRho < MIN_RHO)? MIN_RHO : CLIRho-scaledDeltaRho;
+			A_RhoChanges++;
+			numDecRho++;
+#if ACS_VERBOSE >= VERBOSE
+			PRINT_INFO("Proc: %3d [%s] - FixPolicy::dynamicAdjustRhoMT - Rho Decreased [%5.4f]", threadID, type, CLIRho);
+#endif
+			break;
+
+
+		case CPXMIP_DETTIME_LIM_FEAS:
+		case CPXMIP_TIME_LIM_FEAS:
+		CLIRho = (CLIRho+scaledDeltaRho > MAX_RHO)? MAX_RHO : CLIRho+scaledDeltaRho;
+			A_RhoChanges++;
+			numIncRho++;
+#if ACS_VERBOSE >= VERBOSE
+			PRINT_INFO("Proc: %3d [%s] - FixPolicy::dynamicAdjustRhoMT - Rho Increased [%5.4f]", threadID, type, CLIRho);
+#endif
+			break;
+
+		default: 
+#if ACS_VERBOSE >= VERBOSE
+		PRINT_ERR("Unexpected value for solvecode: %d",solveCode);
+#endif
+
+			break;
+	}
+
+	//Policy in the coin-flip case: make the problem harder
+	if(A_RhoChanges == numMIPs && numIncRho==numDecRho){
+		CLIRho = (CLIRho-DELTA_RHO < MIN_RHO)? MIN_RHO : CLIRho-DELTA_RHO;
+#if ACS_VERBOSE >= VERBOSE
+		PRINT_WARN("[\"COINFLIP\" CASE] - FixPolicy::dynamicAdjustRhoMT - Rho decreased [%5.4f]", CLIRho);
+#endif	
+	} 
+
 }