+++ code/source/MTContext.cpp
@@ -5,6 +5,7 @@ MTContext::MTContext(size_t subMIPNum, unsigned long long intialSeed) : numMIPs{
 	bestACSIncumbent = { .sol = std::vector<double>(), .slackSum = CPX_INFBOUND, .oMIPCost = CPX_INFBOUND };
 	tmpSolutions = std::vector<Solution>();
 	threads = std::vector<std::thread>();
+	A_RhoChanges = 0;
 
 	threads.reserve(numMIPs);
 	tmpSolutions.reserve(numMIPs);
@@ -23,10 +24,11 @@ MTContext::MTContext(size_t subMIPNum, unsigned long long intialSeed) : numMIPs{
 
 void MTContext::setBestACSIncumbent(Solution& sol) {
 
-	if((sol.oMIPCost < bestACSIncumbent.oMIPCost) /*&& (abs(bestACSIncumbent.slackSum ) - abs(sol.slackSum)) <= EPSILON)*/ || abs(sol.slackSum) < abs(bestACSIncumbent.slackSum)){
+	if(((sol.oMIPCost < bestACSIncumbent.oMIPCost) && abs(bestACSIncumbent.slackSum ) - abs(sol.slackSum) <= EPSILON) || abs(sol.slackSum) < abs(bestACSIncumbent.slackSum)){
 
-		std::lock_guard<std::mutex> lock(updateSolMTX);
+		std::lock_guard<std::mutex> lock(MTContextMTX);
 		bestACSIncumbent = { .sol = sol.sol, .slackSum = sol.slackSum, .oMIPCost = sol.oMIPCost };
+		A_RhoChanges = numMIPs;
 
 		if(bestACSIncumbent.oMIPCost< CPX_INFBOUND && bestACSIncumbent.slackSum<= EPSILON)
 			PRINT_BEST("New MIP Incumbent found %12.2f\t[*]", bestACSIncumbent.oMIPCost);
@@ -59,25 +61,26 @@ void MTContext::waitAllJobs() {
 		}
 	}
 	threads.clear();
+	A_RhoChanges=0;
 }
 
 
-MTContext& MTContext::parallelFMIPOptimization(double remTime, Args CLIArgs) {
+MTContext& MTContext::parallelFMIPOptimization(Args& CLIArgs) {
 	waitAllJobs();
 
 	for (size_t i{ 0 }; i < CLIArgs.numsubMIPs; i++) {
-		threads.emplace_back(&MTContext::FMIPInstanceJob, this, i, CLIArgs);
+		threads.emplace_back(&MTContext::FMIPInstanceJob, this, i, std::ref(CLIArgs));
 	}
 
 	waitAllJobs();
 	return *this;
 }
 
-MTContext& MTContext::parallelOMIPOptimization(double remTime, Args CLIArgs, double slackSumUB) {
+MTContext& MTContext::parallelOMIPOptimization(double slackSumUB, Args& CLIArgs) {
 	waitAllJobs();
 
 	for (size_t i{ 0 }; i < CLIArgs.numsubMIPs; i++) {
-		threads.emplace_back(&MTContext::OMIPInstanceJob, this, i, slackSumUB, CLIArgs);
+		threads.emplace_back(&MTContext::OMIPInstanceJob, this, i, slackSumUB, std::ref(CLIArgs));
 	}
 
 	waitAllJobs();
@@ -91,7 +94,7 @@ MTContext::~MTContext() {
 }
 
 
-void MTContext::FMIPInstanceJob(size_t thID, Args CLIArgs) {
+void MTContext::FMIPInstanceJob(const size_t thID, Args& CLIArgs) {
 
 	FMIP fMIP{ CLIArgs.fileName };
 	if (bestACSIncumbent.slackSum < CPX_INFBOUND) {
@@ -99,10 +102,18 @@ void MTContext::FMIPInstanceJob(size_t thID, Args CLIArgs) {
 	}
 	fMIP.setNumCores(CPLEX_CORE);
 
-	FixPolicy::randomRhoFix(tmpSolutions[thID].sol, fMIP, thID, CLIArgs.rho, "FMIP", rndGens[thID]);
+	FixPolicy::randomRhoFixMT(thID, "FMIP", fMIP, tmpSolutions[thID].sol, CLIArgs.rho, rndGens[thID]);
+
+	/// FIXED: Bug #e15760bcfd3dcca51cf9ea23f70072dd6cb2ac14 — Resolved MIPException::WrongTimeLimit triggered by a negligible time limit.
+	if (Clock::timeRemaining(CLIArgs.timeLimit) < EPSILON) {
+#if ACS_VERBOSE >= VERBOSE
+		PRINT_INFO("TIME_LIMIT REACHED");
+#endif
+		return;
+	}
 
 	int solveCode{ fMIP.solve(Clock::timeRemaining(CLIArgs.timeLimit), DET_TL(fMIP.getNumNonZeros())) };
-	if (solveCode == CPXMIP_TIME_LIM_INFEAS || solveCode == CPXMIP_DETTIME_LIM_INFEAS || solveCode == CPXMIP_INFEASIBLE){
+	if (MIP::isINForUNBD(solveCode)){
 #if ACS_VERBOSE >= VERBOSE
 		PRINT_INFO("Proc: %3d [FMIP] - Aborted: Infeasible with given TL", thID);
 #endif
@@ -114,10 +125,12 @@ void MTContext::FMIPInstanceJob(size_t thID, Args CLIArgs) {
 
 	PRINT_OUT("Proc: %3d - FeasMIP Objective: %20.2f", thID, tmpSolutions[thID].slackSum);
 	setBestACSIncumbent(tmpSolutions[thID]);
+
+	FixPolicy::dynamicAdjustRhoMT(thID,"FMIP",solveCode,numMIPs,CLIArgs.rho,A_RhoChanges);
 }
 
 
-void MTContext::OMIPInstanceJob(size_t thID,  double slackSumUB, Args CLIArgs) {
+void MTContext::OMIPInstanceJob(const size_t thID, const double slackSumUB, Args& CLIArgs) {
 
 	OMIP oMIP{ CLIArgs.fileName };
 	if (bestACSIncumbent.slackSum < CPX_INFBOUND) {
@@ -126,10 +139,18 @@ void MTContext::OMIPInstanceJob(size_t thID,  double slackSumUB, Args CLIArgs) {
 	oMIP.setNumCores(CPLEX_CORE);
 	oMIP.updateBudgetConstr(slackSumUB);
 
-	FixPolicy::randomRhoFix(tmpSolutions[thID].sol, oMIP, thID, CLIArgs.rho, "OMIP", rndGens[thID]);
+	FixPolicy::randomRhoFixMT(thID, "OMIP", oMIP, tmpSolutions[thID].sol, CLIArgs.rho, rndGens[thID]);
+	
+	/// FIXED: Bug #e15760bcfd3dcca51cf9ea23f70072dd6cb2ac14 — Resolved MIPException::WrongTimeLimit triggered by a negligible time limit.
+	if (Clock::timeRemaining(CLIArgs.timeLimit) < EPSILON) {
+#if ACS_VERBOSE >= VERBOSE
+		PRINT_INFO("TIME_LIMIT REACHED");
+#endif
+		return;
+	}
 
 	int solveCode{ oMIP.solve(Clock::timeRemaining(CLIArgs.timeLimit), DET_TL(oMIP.getNumNonZeros())) };
-	if (solveCode == CPXMIP_TIME_LIM_INFEAS || solveCode == CPXMIP_DETTIME_LIM_INFEAS || solveCode == CPXMIP_INFEASIBLE) {
+	if (MIP::isINForUNBD(solveCode)) {
 #if ACS_VERBOSE >= VERBOSE
 		PRINT_INFO("Proc: %3d [OMIP] - Aborted: Infeasible with given TL", thID);
 #endif
@@ -142,4 +163,6 @@ void MTContext::OMIPInstanceJob(size_t thID,  double slackSumUB, Args CLIArgs) {
 
 	PRINT_OUT("Proc: %3d - OptMIP Objective: %20.2f|%-10.2f", thID, tmpSolutions[thID].oMIPCost, tmpSolutions[thID].slackSum);
 	setBestACSIncumbent(tmpSolutions[thID]);
+
+	FixPolicy::dynamicAdjustRhoMT(thID,"OMIP",solveCode,numMIPs,CLIArgs.rho,A_RhoChanges);
 }