+++ code/test/scripts/collect_data.py
@@ -116,8 +116,8 @@ def main():
     outs.sort()
 
     cols = {}
-    for rho in [0.10, 0.25, 0.50, 0.75, 0.9]:
-        for seed in [2120934, 3409212, 240931]:
+    for rho in [0.25,0.5,0.75]:
+        for seed in [10493847, 83274910, 70938475, 98312048, 19283746]:
             col = f"ACS_{rho}_{seed}"
             cols[col] = []
     cols["CPLEX"] = []